// src/api/types/AudiencesContactLanguage.ts
//  This file was auto-generated from our API Definition.

/** The contact's detected language. Empty string when no language has been detected or set. */
export const AudiencesContactLanguage = {
    Empty: "",
    En: "en",
    Ar: "ar",
    Af: "af",
    Be: "be",
    Bg: "bg",
    Ca: "ca",
    Zh: "zh",
    ZhCn: "zh_CN",
    Hr: "hr",
    Cs: "cs",
    Da: "da",
    Nl: "nl",
    Et: "et",
    Fa: "fa",
    Fi: "fi",
    Fr: "fr",
    FrCa: "fr_CA",
    De: "de",
    El: "el",
    He: "he",
    Hi: "hi",
    Hu: "hu",
    Is: "is",
    Id: "id",
    Ga: "ga",
    It: "it",
    Ja: "ja",
    Km: "km",
    Ko: "ko",
    Lv: "lv",
    Lt: "lt",
    Mt: "mt",
    Ms: "ms",
    Mk: "mk",
    No: "no",
    Pl: "pl",
    Pt: "pt",
    PtPt: "pt_PT",
    Ro: "ro",
    Ru: "ru",
    Sr: "sr",
    Sk: "sk",
    Sl: "sl",
    Es: "es",
    EsEs: "es_ES",
    Sw: "sw",
    Sv: "sv",
    Ta: "ta",
    Th: "th",
    Tr: "tr",
    Uk: "uk",
    Vi: "vi",
} as const;
export type AudiencesContactLanguage = (typeof AudiencesContactLanguage)[keyof typeof AudiencesContactLanguage];